[feat] build server to delete failed models

+++ builder/utils/deployer.py
@@ -20,26 +20,17 @@ from utils.logging import logger
 
 
 class ModelDeployer:
-    def __init__(self, model_name, s3_uri):
+    def __init__(self, model_name, endpoint_name):
         self.name = model_name
-        s3_bucket, s3_path, self.s3_dir, self.unique_name = self.parse_s3_uri(s3_uri)
+        self.unique_name = endpoint_name
         self.repository_name = self.unique_name.lower()
         self.rootp = tempfile.TemporaryDirectory(prefix=self.unique_name)
         self.root_dir = self.rootp.name
         logger.info(f"{model_name} folder loaded temporarily at {self.root_dir}")
         self.archive_name = f"archive.{self.unique_name}"
-
-        self.env = self.setup_sagemaker_env()
         self.owd = os.getcwd()
 
-        try:
-            self.config_handler = self.initialize(s3_bucket, s3_path)
-            self.config = self.config_handler.load_config()
-        except Exception as ex:
-            logger.exception(
-                f"Exception in fetching model folder and loading config {ex}"
-            )
-            raise RuntimeError("Exception in fetching model folder and loading config")
+        self.env = self.setup_sagemaker_env()
 
     def parse_s3_uri(self, s3_uri):
         parts = s3_uri.replace("s3://", "").split("/")
@@ -49,16 +40,27 @@ class ModelDeployer:
         unique_name = os.path.basename(s3_path).replace(".tar.gz", "")
         return s3_bucket, s3_path, s3_dir, unique_name
 
-    def initialize(self, s3_bucket, s3_path):
-        if os.path.exists(self.root_dir):
-            shutil.rmtree(self.root_dir)
-        os.makedirs(self.root_dir)
-        os.chdir(self.root_dir)
-        logger.info(f"Fetching model folder {s3_path} for {self.name}")
-        self._fetch_folder(s3_bucket, s3_path)
-        logger.info("Load the model setup config")
-        config_handler = SetupConfigHandler(self.name)
-        return config_handler
+    def load_model(self, s3_uri):
+        try:
+            s3_bucket, s3_path, s3_dir, unique_name = self.parse_s3_uri(s3_uri)
+            assert unique_name == self.unique_name, f"Model at S3 path mismatches model endpoint name"
+            if os.path.exists(self.root_dir):
+                shutil.rmtree(self.root_dir)
+            os.makedirs(self.root_dir)
+            logger.info(f"Fetching model folder {s3_path} for {self.name}")
+            self._fetch_folder(s3_bucket, s3_path)
+            logger.info("Load the model setup config")
+            config_handler = SetupConfigHandler(self.name)
+            config = config_handler.load_config()
+            return config_handler, config, s3_dir
+        except AssertionError as ex:
+            logger.exception(ex)
+            raise RuntimeError(ex)
+        except Exception as ex:
+            logger.exception(
+                f"Exception in fetching model folder and loading config {ex}"
+            )
+            raise RuntimeError("Exception in fetching model folder and loading config")
 
     def _fetch_folder(self, s3_bucket, s3_path):
         save_tarball = f"{self.unique_name}.tar.gz"
@@ -71,24 +73,26 @@ class ModelDeployer:
 
     def setup_sagemaker_env(self):
         env = {}
+        try:
+            session = boto3.Session(
+                aws_access_key_id=deploy_config["aws_access_key_id"],
+                aws_secret_access_key=deploy_config["aws_secret_access_key"],
+                region_name=deploy_config["aws_region"],
+            )
+            env["region"] = session.region_name
+            env["account"] = session.client("sts").get_caller_identity().get("Account")
+            env["sagemaker_client"] = session.client("sagemaker")
+            env["s3_client"] = session.client("s3")
+            env["ecr_client"] = session.client("ecr")
 
-        session = boto3.Session(
-            aws_access_key_id=deploy_config["aws_access_key_id"],
-            aws_secret_access_key=deploy_config["aws_secret_access_key"],
-            region_name=deploy_config["aws_region"],
-        )
-        env["region"] = session.region_name
-        env["account"] = session.client("sts").get_caller_identity().get("Account")
-        env["sagemaker_client"] = session.client("sagemaker")
-        env["s3_client"] = session.client("s3")
-        env["ecr_client"] = session.client("ecr")
+            env["sagemaker_session"] = sagemaker.Session(boto_session=session)
+            env["bucket_name"] = env["sagemaker_session"].default_bucket()
 
-        env["sagemaker_session"] = sagemaker.Session(boto_session=session)
-        env["bucket_name"] = env["sagemaker_session"].default_bucket()
+            env["ecr_registry"] = f"{env['account']}.dkr.ecr.{env['region']}.amazonaws.com"
 
-        env["ecr_registry"] = f"{env['account']}.dkr.ecr.{env['region']}.amazonaws.com"
-
-        return env
+            return env
+        except Exception as ex:
+            logger.exception(f"Error in setting up sagemaker environment for deploying model {self.name}")
 
     def delete_existing_endpoints(self):
         # remove endpoint
@@ -149,12 +153,12 @@ class ModelDeployer:
             if response:
                 logger.info(f"Response from deleting ECR repository {response}")
 
-    def build_docker(self, secret):
+    def build_docker(self, secret, setup_config_handler, setup_config):
         # tarball current folder but exclude checkpoints
-        tmp_dir = os.path.join(self.config_handler.config_dir, "tmp")
+        tmp_dir = os.path.join(setup_config_handler.config_dir, "tmp")
         os.makedirs(tmp_dir, exist_ok=True)
         exclude_list_file = os.path.join(tmp_dir, "exclude.txt")
-        self.config_handler.write_exclude_filelist(
+        config_handler.write_exclude_filelist(
             exclude_list_file, self.name, exclude_model=True
         )
         tarball = os.path.join(tmp_dir, f"{self.unique_name}.tar.gz")
@@ -181,7 +185,7 @@ class ModelDeployer:
             shutil.copyfile(os.path.join(docker_dir, f), os.path.join(self.root_dir, f))
 
         # build docker
-        docker_build_args = f"--build-arg tarball={shlex.quote(tarball)} --build-arg requirements={shlex.quote(str(self.config['requirements']))} --build-arg setup={shlex.quote(str(self.config['setup']))} --build-arg my_secret={secret}"
+        docker_build_args = f"--build-arg tarball={shlex.quote(tarball)} --build-arg requirements={shlex.quote(str(setup_config['requirements']))} --build-arg setup={shlex.quote(str(setup_config['setup']))} --build-arg my_secret={secret}"
         docker_build_command = f"docker build --network host -t {shlex.quote(self.repository_name)} -f Dockerfile {docker_build_args} ."
         with subprocess.Popen(
             shlex.split(docker_build_command),
@@ -207,9 +211,9 @@ class ModelDeployer:
 
         os.remove(tarball)
 
-    def build_and_push_docker(self, secret):
+    def build_and_push_docker(self, secret, setup_config_handler, setup_config):
         logger.info(f"Building docker for model {self.name}")
-        self.build_docker(secret)
+        self.build_docker(secret, setup_config_handler, setup_config)
 
         # docker login
         docker_credentials = self.env["ecr_client"].get_authorization_token()[
@@ -271,7 +275,7 @@ class ModelDeployer:
             logger.debug(process.stdout)
         return image_ecr_path
 
-    def archive_and_upload_model(self):
+    def archive_and_upload_model(self, setup_config, s3_dir):
         # torchserve archive model to .tar.gz (.mar)
         # TODO: allow proper model versioning together with docker tag
         logger.info(f"Archiving the model {self.name} ...")
@@ -280,15 +284,15 @@ class ModelDeployer:
             "--model-name",
             shlex.quote(self.archive_name),
             "--serialized-file",
-            shlex.quote(self.config["checkpoint"]),
+            shlex.quote(setup_config["checkpoint"]),
             "--handler",
-            shlex.quote(self.config["handler"]),
+            shlex.quote(setup_config["handler"]),
             "--version",
             "1.0",  # TODO: proper versioning
             "-f",
         ]
-        if self.config["model_files"]:
-            extra_files = ",".join(shlex.quote(f) for f in self.config["model_files"])
+        if setup_config["model_files"]:
+            extra_files = ",".join(shlex.quote(f) for f in setup_config["model_files"])
             archive_command += ["--extra-files", extra_files]
         process = subprocess.run(
             archive_command,
@@ -322,11 +326,11 @@ class ModelDeployer:
         # upload model tarball to S3
         logger.info(f"Uploading the archived model {self.name} to S3 ...")
         response = self.env["s3_client"].upload_file(
-            tarball, self.env["bucket_name"], f"{self.s3_dir}/{tarball}"
+            tarball, self.env["bucket_name"], f"{s3_dir}/{tarball}"
         )
         if response:
             logger.debug(f"Response from the mar file upload to s3 {response}")
-        model_s3_path = f"s3://{self.env['bucket_name']}/{self.s3_dir}/{tarball}"
+        model_s3_path = f"s3://{self.env['bucket_name']}/{s3_dir}/{tarball}"
         os.remove(tarball)
         os.remove(mar)
         return model_s3_path
@@ -350,16 +354,19 @@ class ModelDeployer:
         )
         return f"{deploy_config['gateway_url']}?model={self.unique_name}"
 
-    def deploy(self, secret):
+    def deploy(self, secret, s3_uri):
         self.delete_existing_endpoints()
-        image_ecr_path = self.build_and_push_docker(secret)
-        model_s3_path = self.archive_and_upload_model()
+        os.chdir(self.root_dir)
+        setup_config_handler, setup_config, s3_dir = self.load_model(s3_uri)
+        image_ecr_path = self.build_and_push_docker(secret, setup_config_handler, setup_config)
+        model_s3_path = self.archive_and_upload_model(setup_config, s3_dir)
         endpoint_url = self.deploy_model(image_ecr_path, model_s3_path)
+        self.chdir(self.owd)
         return endpoint_url
 
     def cleanup_post_deployment(self):
+        os.chdir(self.owd)
         try:
-            os.chdir(self.owd)
             self.rootp.cleanup()
             # clean up local docker images
 
@@ -373,14 +380,17 @@ class ModelDeployer:
                 f"Clean up post deployment for {self.unique_name} failed: {ex}"
             )
 
-    def cleanup_on_failure(self):
+    def cleanup_on_failure(self, s3_uri=None):
         try:
             self.cleanup_post_deployment()
             self.delete_existing_endpoints()
-            self.env["s3_client"].delete_object(
-                Bucket=self.env["bucket_name"],
-                Key=f"{self.s3_dir}/{self.archive_name}.tar.gz",
-            )
+            if s3_uri:
+                _, _, s3_dir, unique_name = self.parse_s3_uri(s3_uri)
+                if unique_name == self.unique_name or unique_name == self.archive_name:
+                    self.env["s3_client"].delete_object(
+                        Bucket=self.env["bucket_name"],
+                        Key=f"{s3_dir}/{self.archive_name}.tar.gz",
+                    )
         except Exception as ex:
             logger.exception(
                 f"Clean up on failed deployment for {self.unique_name} failed: {ex}"
